requirements.txt
discord.py==1.4.1
PyNaCl==1.3.0
ffmpeg
discord.py[voice]
youtube-dl
dnspython==1.16.0
cffi==1.14.0
wheel
NHentai-API